resolve use map data util changes
