+++ packages/tupaia-web/src/features/Map/MapLegend/MapLegend.tsx
@@ -8,7 +8,7 @@ import { Legend, LegendProps } from '@tupaia/ui-map-components';
 import { MobileMapLegend } from './MobileMapLegend';
 import { useSearchParams } from 'react-router-dom';
 import { MOBILE_BREAKPOINT, URL_SEARCH_PARAMS } from '../../../constants';
-import { useActiveMapOverlayReport } from '../utils';
+import { useMapOverlayData } from '../utils';
 import styled from 'styled-components';
 
 const DesktopWrapper = styled.div`
@@ -35,7 +35,7 @@ const SeriesDivider = styled.div`
 export const MapLegend = ({ hiddenValues, setValueHidden }: LegendProps) => {
   const [urlSearchParams] = useSearchParams();
   const selectedOverlay = urlSearchParams.get(URL_SEARCH_PARAMS.MAP_OVERLAY);
-  const overlayReportData = useActiveMapOverlayReport();
+  const overlayReportData = useMapOverlayData();
 
   if (!selectedOverlay || !overlayReportData) {
     return null;
