+++ packages/tupaia-web/src/features/Map/MapOverlaySelector/MapOverlayDatePicker.tsx
@@ -7,7 +7,7 @@ import styled from 'styled-components';
 import { useParams } from 'react-router-dom';
 import { Skeleton } from '@material-ui/lab';
 import { useMapOverlays } from '../../../api/queries';
-import { useActiveMapOverlayReport } from '../utils';
+import { useMapOverlayData } from '../utils';
 import { DateRangePicker } from '../../../components';
 import { useDateRanges } from '../../../utils';
 import { URL_SEARCH_PARAMS } from '../../../constants';
@@ -32,7 +32,7 @@ export const MapOverlayDatePicker = () => {
     periodGranularity,
   } = useDateRanges(URL_SEARCH_PARAMS.MAP_OVERLAY_PERIOD, selectedOverlay);
 
-  const { isLoading: isLoadingMapOverlayData } = useActiveMapOverlayReport();
+  const { isLoading: isLoadingMapOverlayData } = useMapOverlayData();
 
   if (!showDatePicker) return null;
   return (
