+++ packages/tupaia-web/src/features/Map/utils/index.ts
@@ -5,6 +5,4 @@
 
 export { useDefaultMapOverlay } from './useDefaultMapOverlay';
 export { useHiddenMapValues } from './useHiddenMapValues';
-export { useMapOverlayReport } from './useMapOverlayReport';
 export { useMapOverlayData } from './useMapOverlayData';
-
